docs: add README.md file
This commit updates the README.md file to provide instructions
and details about the Vagrantfiles-libvirt repository.
It includes a list of prerequisites for using the repository
and a quick start guide for setting up and using the virtual machine.
Additionally, the README.md file explains how to customize the virtual
machine and provides information about contributing to the repository.