feat(vagrant): add archlinux vagrantfiles
This commit addsi an Archlinux Vagrant configuration files
and a playbook to provision the Vagrant box. The Vagrant
configuration sets up a synced folder, specifies a box,
and sets provider-specific configurations.
The playbook installs default packages, Packer,
Terraform, Terragrunt, Sonar-scanner, Git-lfs, and Ohmyzsh
with P10k theme. These tools are commonly used in software
development and will be useful for developers working
in the Vagrant environment.