feat(ansible): disable verbose mode
This commit introduces a new feature to the Ansible configuration by disabling the verbose mode. By removing the verbose output, the execution of Ansible playbooks becomes more streamlined and concise. This enhancement will greatly improve the readability of the logs and make it easier to focus on relevant information during the deployment process.